move some files into sub utils folders
