+++ lib/swid/ir/backend/dart/util/produceDartTranslationUnitsFromBarrelSpec.dart
