+++ lib/swid/ir/backend/translationUnitProducer.dart
@@ -1,9 +1,9 @@
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 
-import 'package:hydro_sdk/swid/ir/backend/dart/produceDartTranslationUnitFromSwidClass.dart';
+import 'package:hydro_sdk/swid/ir/backend/dart/util/produceDartTranslationUnitFromSwidClass.dart';
 import 'package:hydro_sdk/swid/ir/backend/translationUnit.dart';
-import 'package:hydro_sdk/swid/ir/backend/ts/produceTsTranslationUnitFromSwidClass.dart';
+import 'package:hydro_sdk/swid/ir/backend/ts/util/produceTsTranslationUnitFromSwidClass.dart';
 import 'package:hydro_sdk/swid/ir/backend/ts/tsEnum.dart';
 import 'package:hydro_sdk/swid/ir/backend/ts/tsTranslationUnit.dart';
 import 'package:hydro_sdk/swid/ir/backend/ts/tsir.dart';
