+++ lib/swid/ir/backend/ts/util/produceTsTranslationUnitFromSwidClass.dart
